requirements.txt
streamlit
faiss-cpu
sentence-transformers
python-pptx
groq
rank-bm25